scheduler/requirements.txt
kafka-python==2.0.2
kubernetes==30.1.0
PyYAML==5.4.1
requests==2.31.0

